# ios/SplitBackAPI/Sources/SplitBackAPI/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SplitBack",
    "version": "0.1.0"
  },
  "paths": {
...
    "/expenses": {
      "post": {
        "tags": [
          "expenses"
        ],
        "summary": "Create Expense",
        "operationId": "create_expense_expenses_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExpenseCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExpenseResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "expenses"
        ],
        "summary": "List Expenses",
        "operationId": "list_expenses_expenses_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "group_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Group Id"
            }
          },
          {
            "name": "since",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date",
              "title": "Since"
            }
          },
          {
            "name": "until",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date",
              "title": "Until"
            }
          },
          {
            "name": "updated_since",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date-time",
              "title": "Updated Since"
            }
          },
          {
            "name": "include_archived",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Include Archived"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ExpenseResponse"
                  },
                  "title": "Response List Expenses Expenses Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/expenses/{expense_id}": {
      "get": {
        "tags": [
          "expenses"
        ],
        "summary": "Get Expense",
        "operationId": "get_expense_expenses__expense_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "expense_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Expense Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExpenseResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "expenses"
        ],
        "summary": "Update Expense",
        "operationId": "update_expense_expenses__expense_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "expense_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Expense Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExpenseUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExpenseResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "expenses"
        ],
        "summary": "Delete Expense",
        "operationId": "delete_expense_expenses__expense_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "expense_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Expense Id"
            }
          },
          {
            "name": "propagate",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "title": "Propagate"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ExpenseCreate": {
        "properties": {
          "group_id": {
            "type": "string",
            "format": "uuid",
            "title": "Group Id"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "amount": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Amount"
          },
          "currency": {
            "type": "string",
            "title": "Currency"
          },
          "date": {
            "type": "string",
            "format": "date",
            "title": "Date"
          },
          "category": {
            "type": "string",
            "title": "Category"
          },
          "transaction_id": {
            "type": "string",
            "format": "uuid",
            "title": "Transaction Id"
          },
          "splits": {
            "items": {
              "$ref": "#/components/schemas/SplitInput"
            },
            "type": "array",
            "title": "Splits",
            "default": []
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/ItemInput"
            },
            "type": "array",
            "title": "Items",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "group_id",
          "description",
          "amount",
          "date"
        ],
        "title": "ExpenseCreate"
      },
      "ExpenseResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "group_id": {
            "type": "string",
            "format": "uuid",
            "title": "Group Id"
          },
          "transaction_id": {
            "type": "string",
            "format": "uuid",
            "title": "Transaction Id"
          },
          "splitwise_expense_id": {
            "type": "string",
            "title": "Splitwise Expense Id"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "amount": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Amount"
          },
          "currency": {
            "type": "string",
            "title": "Currency"
          },
          "date": {
            "type": "string",
            "format": "date",
            "title": "Date"
          },
          "category": {
            "type": "string",
            "title": "Category"
          },
          "archived_at": {
            "type": "string",
            "format": "date-time",
            "title": "Archived At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "splits": {
            "items": {
              "$ref": "#/components/schemas/SplitResponse"
            },
            "type": "array",
            "title": "Splits",
            "default": []
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/ItemResponse"
            },
            "type": "array",
            "title": "Items",
            "default": []
          },
          "receipts": {
            "items": {
              "$ref": "#/components/schemas/ReceiptResponse"
            },
            "type": "array",
            "title": "Receipts",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "id",
          "group_id",
          "description",
          "amount",
          "currency",
          "date",
          "created_at",
          "updated_at"
        ],
        "title": "ExpenseResponse"
      },
      "ExpenseUpdate": {
        "properties": {
          "group_id": {
            "type": "string",
            "format": "uuid",
            "title": "Group Id"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "amount": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Amount"
          },
          "currency": {
            "type": "string",
            "title": "Currency"
          },
          "date": {
            "type": "string",
            "format": "date",
            "title": "Date"
          },
          "category": {
            "type": "string",
            "title": "Category"
          },
          "transaction_id": {
            "type": "string",
            "format": "uuid",
            "title": "Transaction Id"
          },
          "splits": {
            "items": {
              "$ref": "#/components/schemas/SplitInput"
            },
            "type": "array",
            "title": "Splits"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/ItemInput"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "title": "ExpenseUpdate"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ItemInput": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "quantity": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Quantity",
            "default": "1"
          },
          "price": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Price"
          },
          "category": {
            "type": "string",
            "title": "Category"
          }
        },
        "type": "object",
        "required": [
          "name",
          "price"
        ],
        "title": "ItemInput"
      },
      "ItemResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "quantity": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Quantity"
          },
          "price": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Price"
          },
          "category": {
            "type": "string",
            "title": "Category"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "quantity",
          "price"
        ],
        "title": "ItemResponse"
      },
...
      "ReceiptResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "expense_id": {
            "type": "string",
            "format": "uuid",
            "title": "Expense Id"
          },
          "bucket": {
            "type": "string",
            "title": "Bucket"
          },
          "object_key": {
            "type": "string",
            "title": "Object Key"
          },
          "content_type": {
            "type": "string",
            "title": "Content Type"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "expense_id",
          "bucket",
          "object_key",
          "created_at"
        ],
        "title": "ReceiptResponse"
      },
      "SplitInput": {
        "properties": {
          "user_identifier": {
            "type": "string",
            "title": "User Identifier"
          },
          "paid_share": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Paid Share"
          },
          "owed_share": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Owed Share"
          }
        },
        "type": "object",
        "required": [
          "user_identifier",
          "paid_share",
          "owed_share"
        ],
        "title": "SplitInput"
      },
      "SplitResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "user_identifier": {
            "type": "string",
            "title": "User Identifier"
          },
          "paid_share": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Paid Share"
          },
          "owed_share": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Owed Share"
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_identifier",
          "paid_share",
          "owed_share"
        ],
        "title": "SplitResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}